% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multi_rev_per_session.R
\name{calculate_multi_rev_per_session}
\alias{calculate_multi_rev_per_session}
\title{Calculate Multi Rev Per Session}
\usage{
calculate_multi_rev_per_session(conv_rates, inverse_rev_A, inverse_rev_B)
}
\arguments{
\item{conv_rates}{Dirichlet samples containing a tibble with columns alpha_1, alpha_2, and alpha_0}

\item{inverse_rev_A}{Vector of inverse revenue samples from A conversion type}

\item{inverse_rev_B}{Vector of inverse revenue samples from B conversion type}
}
\value{
Vector of samples (dbl)
}
\description{
Calculate Multi Rev Per Session
}
