% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_total_cm.R
\name{calculate_total_cm}
\alias{calculate_total_cm}
\title{Calculate Total CM}
\usage{
calculate_total_cm(rev_per_click, cost_per_click, expected_clicks)
}
\arguments{
\item{rev_per_click}{vector of rev per click samples}

\item{cost_per_click}{vector of cost per click (cpc) samples}

\item{expected_clicks}{vector of expected clicks (expected CTR * fixed impressions)}
}
\value{
vector of CM estimates (dbl)
}
\description{
Calculate Total CM
}
