% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{distribution_column_mapping}
\alias{distribution_column_mapping}
\title{A Mapping from distribution names (inputs) to columns. This allows the
package to dynamically select column names for different distribution types.}
\format{
A data frame with 11 rows and 12 columns
}
\usage{
distribution_column_mapping
}
\description{
A Mapping from distribution names (inputs) to columns. This allows the
package to dynamically select column names for different distribution types.
}
\keyword{datasets}
