% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_option.R
\name{find_best_option}
\alias{find_best_option}
\title{Find Best Option}
\usage{
find_best_option(posterior_samples, distribution)
}
\arguments{
\item{posterior_samples}{Tibble returned from sample_from_posterior with 3 columns
`option_name`, `samples`, and `sample_id`.}

\item{distribution}{String: name of the distribution}
}
\value{
String: the best option name
}
\description{
Samples from posterior, calculates win probability, and selects the best option.
Note: this can be inefficient if you already have the win probability dataframe.
Only use this if that has not already been calculated.
}
\examples{
# Requires posterior distribution
\dontrun{
find_best_option(posterior_samples = posterior_samples, distribution = "conversion_rate")
}
}
