% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grizbayr-package.R
\docType{package}
\name{grizbayr-package}
\alias{grizbayr}
\alias{grizbayr-package}
\title{grizbayr: Bayesian Inference for A|B and Bandit Marketing Tests}
\description{
Uses simple Bayesian conjugate prior update rules to calculate the win probability of each option, value remaining in the test, and percent lift over the baseline for various marketing objectives. References: Fink, Daniel (1997) "A Compendium of Conjugate Priors" \url{https://www.johndcook.com/CompendiumOfConjugatePriors.pdf}. Stucchio, Chris (2015) "Bayesian A/B Testing at VWO" \url{https://vwo.com/downloads/VWO_SmartStats_technical_whitepaper.pdf}.
}
\keyword{internal}
