% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grizbayr.R
\docType{package}
\name{grizbayr}
\alias{grizbayr}
\title{\code{grizbayr} package}
\description{
Grizzly Bear - Bayesian Inference Package for A|B and Bandit Marketing Tests
}
\details{
See the README on
\href{https://github.com/rangi513/grizbayr/blob/master/README.md}{GitHub}
or the `intro` vignette contained in the package.
}
