% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_winner_max.R
\name{is_winner_max}
\alias{is_winner_max}
\title{Is Winner Max}
\usage{
is_winner_max(distribution)
}
\arguments{
\item{is_winner_max}{}
}
\value{
Boolean TRUE/FALSE
}
\description{
Determines if the max or min function should be used for win probability.
If CPA or CPC distribution, lower is better, else higher number is better.
}
