% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_column.R
\name{validate_input_column}
\alias{validate_input_column}
\title{Validate Input Column}
\usage{
validate_input_column(column_name, input_df, greater_than_zero = TRUE)
}
\arguments{
\item{column_name}{String value of the column name}

\item{input_df}{Dataframe containing option_name (str) and various other columns
depending on the distribution type. See vignette for more details.}
}
\value{
None
}
\description{
Validates the input column exists in the dataframe, is of the correct type,
and that all values are greater than or equal to 0.
}
