% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_posterior_samples.R
\name{validate_posterior_samples}
\alias{validate_posterior_samples}
\title{Validate Posterior Samples Dataframe}
\usage{
validate_posterior_samples(posterior_samples)
}
\arguments{
\item{posterior_samples}{Tibble of data in long form with 2 columns
`option_name` and `samples`}
}
\value{
None
}
\description{
Function fails if posterior is not shaped correctly.
}
