% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_options.R
\name{impute_missing_options}
\alias{impute_missing_options}
\title{Impute Missing Options}
\usage{
impute_missing_options(posterior_samples, wp_raw)
}
\arguments{
\item{posterior_samples}{Tibble of data in long form with 2 columns
`option_name` and `samples`}

\item{wp_raw}{Tibble of win probabilities with the columns `option_name` and `win_prob_raw``}
}
\value{
wp_raw table with new rows if opption names were missing.
}
\description{
When win probability is calculated
}
\examples{
post_sample_example <- tibble::tibble(option_name = c("A", "B", "C"))
wp_raw <- tibble::tibble(option_name = c("B", "C"), win_prob_raw = c(0.4, 0.6))
impute_missing_options(posterior_samples = post_sample_example, wp_raw = wp_raw)
}
