% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{update_beta}
\alias{update_beta}
\title{Update Beta}
\usage{
update_beta(successes, failures, priors = list())
}
\arguments{
\item{successes}{Double value of successes. Must be 0 or greater.}

\item{failures}{Double value of failures. Must be 0 or greater.}

\item{priors}{An optional list object that contains alpha0 and
beta0. Otherwise the function with use Beta(1,1) as the prior distribution.}
}
\value{
A list object that contains `alpha` and `beta`
}
\description{
Updates Beta Distribution with the Beta-Bernoulli
conjugate prior update rule
}
\examples{
update_beta(successes = 1, failures = 5, priors = list(alpha0 = 2, beta0 = 2))
update_beta(successes = 20000, failures = 50000)

}
