% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_values.R
\name{validate_data_values}
\alias{validate_data_values}
\title{Validate Data Values}
\usage{
validate_data_values(data_values)
}
\arguments{
\item{data_values}{List of named data values}
}
\value{
None
}
\description{
Validate Data Values
}
\examples{
validate_data_values(data_values = list(successes = 22, failures = 100))
## DONT RUN
\donttest{
validate_data_values(data_values = list(successes = 22, failures = -1))
}

}
