% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_priors.R
\name{validate_priors}
\alias{validate_priors}
\title{Validate Priors}
\usage{
validate_priors(priors, valid_priors, default_priors)
}
\arguments{
\item{priors}{List of named priors with double values.}

\item{valid_priors}{A character vector of valid prior names.}

\item{default_priors}{A list of default priors for the distribution.}
}
\value{
A named list of valid priors for the distribution.
}
\description{
Validates list of priors against a vector of valid priors and if
   the values are not valid, default priors are returned.
}
\examples{
priors <- list(a = 1, b = 2, c = 3)
valid_priors <- c("b", "c")
default_priors <- list(b = 1, c = 1)
validate_priors(priors, valid_priors, default_priors)

}
